Renamed addcol to newcol in PrettySeries.
No major changes in this commit.  Just didn't like the name,
addcol, for defining a new column to PrettySeries.  So, using
newcol...not sure if best name but it'll do for now.

Also, just didn't like passing *values as an argument for
PrettySeries.text or PrettySeries.format.  Most of the time,
users will be passing in a list of something.  So, would rather
build the call to handle what happens most of the time and in
those rare cases when user wants to format a value they can
just encapsulate within a [].