Added ability to print title of text table.
Title of the table can be passed and built
into the resulting text table for the text
method call. Please note the title is truncated
to the max size of the text table row.  Future
plans are in the works to breakup a title that
is too long for one row into multiple rows.